+++ scout/server/blueprints/variants/controllers.py
@@ -165,7 +165,6 @@ def parse_variant(store, institute_obj, case_obj, variant_obj, update=False):
     # convert length for SV variants
     variant_length = variant_obj.get('length')
     variant_obj['length'] = {100000000000: 'inf', -1: 'n.d.'}.get(variant_length, variant_length)
-    pp(variant_obj)
 
     return variant_obj